# api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NFC Loyalty & Customer Growth Platform API",
    "description": "REST API untuk platform loyalty berbasis NFC/QR untuk UMKM Indonesia. Menyediakan fitur registrasi member, pemesanan, validasi pembayaran manual, poin loyalitas, voucher, dan dashboard analytics.",
    "version": "1.0.0",
    "contact": { "name": "NFC Loyalty Platform" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development server" }
  ],
...
  "paths": {
...
    "/api/menu/categories": {
      "get": {
        "tags": ["Menu"],
        "summary": "Get all menu categories with items (public)",
        "operationId": "getMenuCategories",
        "parameters": [
          { "name": "tenantId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "List of categories with menu items",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "cat-001",
                      "tenantId": "550e8400-e29b-41d4-a716-446655440000",
                      "name": "Makanan",
                      "sortOrder": 1,
                      "items": [
                        {
                          "id": "item-001",
                          "categoryId": "cat-001",
                          "name": "Nasi Goreng Spesial",
                          "description": "Nasi goreng dengan telur dan ayam",
                          "price": 25000,
                          "imageUrl": "/uploads/nasi-goreng.jpg",
                          "isAvailable": true,
                          "sortOrder": 1
                        },
                        {
                          "id": "item-002",
                          "categoryId": "cat-001",
                          "name": "Mie Goreng",
                          "description": "Mie goreng spesial",
                          "price": 20000,
                          "imageUrl": "/uploads/mie-goreng.jpg",
                          "isAvailable": true,
                          "sortOrder": 2
                        }
                      ]
                    },
                    {
                      "id": "cat-002",
                      "tenantId": "550e8400-e29b-41d4-a716-446655440000",
                      "name": "Minuman",
                      "sortOrder": 2,
                      "items": [
                        {
                          "id": "item-003",
                          "categoryId": "cat-002",
                          "name": "Es Teh Manis",
                          "description": "Teh manis dingin",
                          "price": 8000,
                          "imageUrl": null,
                          "isAvailable": true,
                          "sortOrder": 1
                        }
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/menu/categories": {
      "post": {
        "tags": ["Menu"],
        "summary": "Create menu category (admin)",
        "operationId": "createMenuCategory",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCategoryRequest" },
              "example": { "name": "Makanan", "sortOrder": 1 }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "d4e5f6a7-b8c9-0123-defg-456789012345",
                    "tenantId": "550e8400-e29b-41d4-a716-446655440000",
                    "name": "Makanan",
                    "sortOrder": 1
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/menu/items": {
      "post": {
        "tags": ["Menu"],
        "summary": "Create menu item (admin)",
        "operationId": "createMenuItem",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateMenuItemRequest" },
              "example": {
                "name": "Nasi Goreng Spesial",
                "description": "Nasi goreng dengan telur, ayam, dan sayuran",
                "price": 25000,
                "categoryId": "d4e5f6a7-b8c9-0123-defg-456789012345",
                "imageUrl": "https://example.com/nasi-goreng.jpg",
                "isAvailable": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Menu item created",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "e5f6a7b8-c9d0-1234-efgh-567890123456",
                    "tenantId": "550e8400-e29b-41d4-a716-446655440000",
                    "categoryId": "d4e5f6a7-b8c9-0123-defg-456789012345",
                    "name": "Nasi Goreng Spesial",
                    "description": "Nasi goreng dengan telur, ayam, dan sayuran",
                    "price": 25000,
                    "imageUrl": "https://example.com/nasi-goreng.jpg",
                    "isAvailable": true,
                    "sortOrder": 0,
                    "createdAt": "2025-07-15T10:30:00.000Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "example": { "success": false, "error": { "message": "Kategori tidak ditemukan" } }
              }
            }
          }
        }
      }
    },
    "/api/admin/menu/items/{id}": {
      "patch": {
        "tags": ["Menu"],
        "summary": "Update menu item (admin)",
        "operationId": "updateMenuItem",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateMenuItemRequest" },
              "example": { "name": "Nasi Goreng Super", "price": 28000, "isAvailable": false }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Menu item updated",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "e5f6a7b8-c9d0-1234-efgh-567890123456",
                    "name": "Nasi Goreng Super",
                    "price": 28000,
                    "isAvailable": false
                  }
                }
              }
            }
          },
          "404": {
            "description": "Menu item not found",
            "content": {
              "application/json": {
                "example": { "success": false, "error": { "message": "Menu item tidak ditemukan" } }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Menu"],
        "summary": "Delete menu item (admin)",
        "operationId": "deleteMenuItem",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Menu item deleted",
            "content": {
              "application/json": {
                "example": { "success": true, "data": { "message": "Menu item berhasil dihapus" } }
              }
            }
          },
          "404": {
            "description": "Menu item not found",
            "content": {
              "application/json": {
                "example": { "success": false, "error": { "message": "Menu item tidak ditemukan" } }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token (15min). Get from /api/auth/register, /api/auth/login, or /api/auth/admin/login"
      }
    },
    "schemas": {
...
      "CreateCategoryRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 50 },
          "sortOrder": { "type": "integer", "default": 0 }
        }
      },
      "CreateMenuItemRequest": {
        "type": "object",
        "required": ["name", "price", "categoryId"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "description": { "type": "string", "maxLength": 200 },
          "price": { "type": "integer", "minimum": 0, "description": "Price in IDR (integer)" },
          "categoryId": { "type": "string", "format": "uuid" },
          "imageUrl": { "type": "string", "format": "uri" },
          "isAvailable": { "type": "boolean", "default": true }
        }
      },
      "UpdateMenuItemRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 100 },
          "description": { "type": "string", "maxLength": 200 },
          "price": { "type": "integer", "minimum": 0 },
          "categoryId": { "type": "string", "format": "uuid" },
          "imageUrl": { "type": "string", "nullable": true },
          "isAvailable": { "type": "boolean" }
        }
      },
...
    }
  }
}